Move mock_kprint to own file for other tests can use them
